Implemented user creation and authentication using JWT tokens
Using Passport.js for authorization (Strategies and JWT)
Implemented:
- Custom param decorator for fetching the current user
- Password hashing using bcrypt
- Middleware for CRUD operations with @UseGuards decorator

+++ src/events/events.controller.ts
@@ -2,6 +2,7 @@ import {
     Body,
     Controller,
     Delete,
+    ForbiddenException,
     Get,
     HttpCode,
     Logger,
@@ -11,28 +12,22 @@ import {
     Patch,
     Post,
     Query,
+    UseGuards,
     UsePipes,
     ValidationPipe,
 } from '@nestjs/common';
 import { CreateEventDTO, UpdateEventDTO } from './input/events.dto';
-import { Event } from './events.entity';
-import { Attendee } from './attendee.entity';
-import { Like, MoreThan, Repository } from 'typeorm';
-import { InjectRepository } from '@nestjs/typeorm';
 import { EventsService } from './events.service';
 import { ListEvents } from './input/list.events';
+import { CurrentUser } from 'src/auth/current-user.decorator';
+import { User } from 'src/auth/user.entity';
+import { AuthGuardJwt } from 'src/auth/auth-guard.jwt';
 
 @Controller('/events')
 export class EventsController {
     private readonly logger = new Logger(EventsController.name);
 
-    constructor(
-        @InjectRepository(Event)
-        private readonly repository: Repository<Event>,
-        @InjectRepository(Attendee)
-        private readonly attendeeRepository: Repository<Attendee>,
-        private readonly eventsService: EventsService,
-    ) {}
+    constructor(private readonly eventsService: EventsService) {}
 
     @Get()
     @UsePipes(new ValidationPipe({ transform: true }))
@@ -49,46 +44,6 @@ export class EventsController {
         return events;
     }
 
-    @Get('/practice')
-    async practice() {
-        return await this.repository.find({
-            select: ['id', 'when'],
-            where: [
-                {
-                    id: MoreThan(3),
-                    when: MoreThan(new Date('2021-02-12T13:00:00')),
-                },
-                {
-                    description: Like('%meet%'),
-                },
-            ],
-            take: 2,
-            order: {
-                id: 'DESC',
-            },
-        });
-    }
-
-    @Get('practice2')
-    async practice2() {
-        // return await this.repository.findOne({
-        //     where: {
-        //         id: 1,
-        //     },
-        //     loadEagerRelations: false,
-        // });
-        const event = new Event();
-        event.id = 1;
-
-        const attendee = new Attendee();
-        attendee.name = 'Jerry The Second';
-        attendee.event = event;
-
-        await this.attendeeRepository.save(attendee);
-
-        return event;
-    }
-
     @Get(':id')
     async findOne(@Param('id', ParseIntPipe) id: number) {
         const event = await this.eventsService.getEvent(id);
@@ -101,38 +56,96 @@ export class EventsController {
     }
 
     @Post()
-    async create(@Body() input: CreateEventDTO) {
-        return await this.repository.save({
-            ...input,
-            when: new Date(input.when),
-        });
+    @UseGuards(AuthGuardJwt)
+    async create(@Body() input: CreateEventDTO, @CurrentUser() user: User) {
+        return await this.eventsService.createEvent(input, user);
     }
 
     @Patch(':id')
+    @UseGuards(AuthGuardJwt)
     async update(
         @Param('id', ParseIntPipe) id: number,
         @Body() input: UpdateEventDTO,
+        @CurrentUser() user: User,
     ) {
-        const event = await this.repository.findOneBy({ id });
+        const event = await this.eventsService.getEvent(id);
 
         if (!event) {
             throw new NotFoundException();
         }
 
-        return await this.repository.save({
-            ...event,
-            ...input,
-            when: input.when ? new Date(input.when) : event.when,
-        });
+        if (event.organizerId !== user.id) {
+            throw new ForbiddenException(
+                null,
+                `You are not authorized to change this event`,
+            );
+        }
+
+        return await this.eventsService.updateEvent(event, input);
     }
 
     @Delete(':id')
+    @UseGuards(AuthGuardJwt)
     @HttpCode(204)
-    async remove(@Param('id', ParseIntPipe) id: number) {
-        const result = await this.eventsService.deleteEvent(id);
+    async remove(
+        @Param('id', ParseIntPipe) id: number,
+        @CurrentUser() user: User,
+    ) {
+        const event = await this.eventsService.getEvent(id);
 
-        if (result.affected !== 1) {
+        if (!event) {
             throw new NotFoundException();
         }
+
+        if (event.organizerId !== user.id) {
+            throw new ForbiddenException(
+                null,
+                `You are not authorized to remove this event`,
+            );
+        }
+
+        await this.eventsService.deleteEvent(id);
     }
+
+    /*
+    @Get('/practice')
+    async practice() {
+        // return await this.repository.find({
+        //     select: ['id', 'when'],
+        //     where: [
+        //         {
+        //             id: MoreThan(3),
+        //             when: MoreThan(new Date('2021-02-12T13:00:00')),
+        //         },
+        //         {
+        //             description: Like('%meet%'),
+        //         },
+        //     ],
+        //     take: 2,
+        //     order: {
+        //         id: 'DESC',
+        //     },
+        // });
+    }
+
+    @Get('practice2')
+    async practice2() {
+        // return await this.repository.findOne({
+        //     where: {
+        //         id: 1,
+        //     },
+        //     loadEagerRelations: false,
+        // });
+        // const event = new Event();
+        // event.id = 1;
+
+        // const attendee = new Attendee();
+        // attendee.name = 'Jerry The Second';
+        // attendee.event = event;
+
+        // await this.attendeeRepository.save(attendee);
+
+        // return event;
+    }
+    */
 }
